+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 ## UNRELEASED
 
-### Adds
+### Changes
+* Updates area UX to not display Add Content controls when a widget is focused
+* Updates area UX to unfocus widget on esc key
+* Updates widget UI to use dashed outlines instead of borders to indicate bounds
+* Updates UI for Insert Menu
+* Updates Insert Menu UX to allow mid-node insertion
+* Rich Text Widget's Insert components are now expected to emit `done` and `cancel` for proper RT cleanup. `close` still supported for BC, acts as `done`.
 
+### Adds
+* Adds keyboard accessibility to Insert menu
 * Adds regex pattern feature for string fields.
 * Adds a warning if database queries involving relationships
 are made before the last `apostrophe:modulesRegistered` handler has fired.